+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 the runtime of Minecord.
 - On Alpha Update 2.0 and 3.0, the plugin is recoded.
 
+## 1.0.3
+- Adding identities now doesn't throw DuplicatedException, and replaced with returning boolean value.
+- Planning to remove ChannelManager
+
 ## 1.0.2
 - Fixed a bug that prints minecord.properties
 - Fixed automatic ServerIdentity generation bug